--- a/.github/actions/setup-java/action.yml
+++ b/.github/actions/setup-java/action.yml
@@ -30,4 +30,5 @@ runs:
       with:
         java-version: '17'
         distribution: 'temurin'
-        cache: 'gradle'
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4