## WHAT

As one of the impactful Eclipse projects, we (the Develocity Solutions team) would like to invite you to be a part of the [Eclipse Develocity evaluation initiative](https://www.eclipse.org/lists/cross-project-issues-dev/msg20000.html).

This improvement will enhance the functionality of the TractusX EDC build by publishing build scans to [develocity-staging.eclipse.org](https://develocity-staging.eclipse.org/), hosted by the Eclipse Foundation and run in partnership between the Eclipse Foundation and Gradle. This Develocity is freely available for use by the Eclipse TractusX EDC project and all other Eclipse projects.

Here is a [blog post abut Develocity](https://eclipse-cbi.github.io/cbi-website/blog/2025/02/12/2025-02-12_develocity) from Eclipse Foundation, as well as a [Develocity presentation](https://www.youtube.com/watch?v=YWIbzmNFJVM), made in collaboration with Eclipse Foundation (and Gradle).

More information can be read in the [Eclipse announcement](https://www.eclipse.org/lists/cross-project-issues-dev/msg20000.html). 

## WHY

On this Develocity instance, the Eclipse TractusX EDC project will have access not only to all of the published build scans but also to other aggregate data features such as:

* Dashboards to view all historical build scans, along with performance [trends](https://develocity-staging.eclipse.org/scans/trends?search.relativeStartTime=P28D&trends.timeResolution=week) over time. For example, look at the [Quarkus instance trends dashboard](https://ge.quarkus.io/scans/trends?search.relativeStartTime=P28D&search.tags=ci&search.timeZoneId=Europe%2FLjubljana) with the `ci` filter applied.
* [Build failure analytics](https://develocity-staging.eclipse.org/scans/failures?search.relativeStartTime=P28D) for enhanced investigation and diagnosis of build failures. You can also explore the [Quarkus instance](https://ge.quarkus.io/scans/failures?search.relativeStartTime=P28D&search.timeZoneId=Europe%2FLjubljana) for example of a different OSS projects Develocity instance.
* [Test failure analytics](https://develocity-staging.eclipse.org/scans/tests?search.relativeStartTime=P28D) to better understand trends and causes around slow, failing, and flaky tests. Going to the [Quarkus instance](https://ge.quarkus.io/scans/tests?search.relativeStartTime=P28D&search.timeZoneId=Europe%2FLjubljana) again, this will probably give you a better representation of the type of data available. You can also order it by flakiness, which is a good starting point when deciding which tests need fixing the most.

This will also enable you to (optionally) use build time optimization features, such as build caching and Predictive Test Selection. 

I ran [some tests](https://github.com/ribafish/kotlinTests/actions/runs/13500918944) about build speed improvements when caching is configured with the following results - there is additional ~13s savings available when the project is relocated, but that would require a fix in the [EDC Gradle Plugins / AutodocDependencyInjector](https://github.com/eclipse-edc/GradlePlugins/blob/60afb05e209307376abeaf88aae947354e84e430/plugins/autodoc/autodoc-plugin/src/main/java/org/eclipse/edc/plugins/autodoc/AutodocDependencyInjector.java#L56). Here are the best case savings:

| Invocation | Build time without cache | Build time with cache | Build time savings |
|-----------|--------------------------|-----------------------|-------------------|
| `assemble` | 4m 33.373s | 56.972s | 3m 36.401s (79%) |
| `build` | 6m 43.965s | 58.916s | 5m 45.049s (85%) |
| `test` | 3m 41.765s | 7.475s | 3m 34.290s (96%) |
| `javadoc` | 2m 28s | 4s | 2m 24s (97%) |

## FURTHER NOTES

**IMPORTANT:** If your project is interested in being a part of this initiative, please reach out to Eclipse infra by filling out a [helpdesk ticket](https://gitlab.eclipse.org/eclipsefdn/helpdesk/-/issues/new) to get the CI setup and credentials.

Closes https://github.com/eclipse-tractusx/tractusx-edc/issues/1815